revert incorrect fix for put updating shadow_index, fixes #5661
A) the compaction code needs the shadow index only for this case:

segment A: PUT x, segment B: DEL x, with A < B  (DEL shadows the PUT).

B) for the following case, we have no shadowing DEL (or rather: it does not matter,
because there is a PUT right after the DEL) and x is in the repo index,
thus the shadow_index is not needed for the special case in the compaction code:

segment A: PUT x, segment B: DEL x PUT x

see also PR #5636.

reverts dc376755be3d70a468b8a07aa4597e1c886fa96c
and clarifies the code by more comments.

we keep the code deduplication of fffd73b6b3147321545711272ce119ebf75c38bd
and just add a update_shadow_index param to make it not look like there was
something accidentally forgotten, which was the whole reason for the reverted
"fix".