Remove --list-format from borg list
While reading the docs I noticed that in `borg list` the options --list-format and --format do the same thing. Using `git log -S` I have uncovered that --list-format used to be deprecated and was supposed to be removed in 738d7336210fec2258886c6a2c79e4f7cfad041a, but you overlooked it and undeprecated it instead. What should we do now? Just remove it or deprecate it again?